feat: add params for the initialize method
